requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
pyannote.audio==3.1.1
googletrans==4.0.0rc1
TTS==0.22.0
torch==2.1.0
torchaudio==2.1.0
moviepy==1.0.3
pydub==0.25.1
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1